# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: StreamFlix API
  version: 1.0.0
  description: |
    REST API for the StreamFlix backend — a Netflix-like streaming service.

    ## Auth
    All endpoints except `POST /auth/register` and `POST /auth/login` require a
    JWT in the `Authorization: Bearer <token>` header. Tokens are issued by the
    login/register endpoints and are valid for 24 hours.

    Video streaming uses a separate short-lived token passed as a query
    parameter (`?token=…`) so that `<video src>` works without a custom header.

    ## Errors
    All non-2xx responses use the [RFC 9457 Problem Details](https://www.rfc-editor.org/rfc/rfc9457.html)
    format. See `components.schemas.Problem`.
  contact:
    name: StreamFlix
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: /api
    description: Same-origin (production)
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users (paginated)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/Order'
        - name: email
          in: query
          description: Case-insensitive substring filter on email
          schema: { type: string, maxLength: 100 }
        - name: isActive
          in: query
          schema: { type: string, enum: ['true', 'false'] }
      responses:
        '200':
          description: A page of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/ObjectId'
    get:
      tags: [Users]
      summary: Get one user
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Users]
      summary: Update mutable user fields
      description: Only `birthDate`, `walletBalance`, and `isActive` can be updated here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                birthDate: { type: string, format: date }
                walletBalance: { type: number, minimum: 0 }
                isActive: { type: boolean }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Users]
      summary: Delete a user (admin only)
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────────────────────────────────────── movies ──
...
  /reviews:
    get:
      tags: [Reviews]
      summary: List reviews (paginated, filterable)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: sortBy
          in: query
          schema: { type: string, enum: [createdAt, rating], default: createdAt }
        - $ref: '#/components/parameters/Order'
        - name: movieId
          in: query
          schema: { type: string }
        - name: userId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: A page of reviews
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Review' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Reviews]
      summary: Create a review
      description: The authenticated user becomes the review's author automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewCreateRequest' }
      responses:
        '201':
          description: Created review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /reviews/{id}:
    parameters:
      - $ref: '#/components/parameters/ObjectId'
    get:
      tags: [Reviews]
      summary: Get one review
      responses:
        '200':
          description: Review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Reviews]
      summary: Update a review
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewUpdateRequest' }
      responses:
        '200':
          description: Updated review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Reviews]
      summary: Delete a review
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

# ─────────────────────────────────────────────────────────── components ──
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by `/auth/login` or `/auth/register`. Valid for 24 hours.
  parameters:
    ObjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
        example: 65f3a1d2c8e7b9001fae12d4
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    SortBy:
      name: sortBy
      in: query
      schema: { type: string, default: createdAt }
    Order:
      name: order
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
  responses:
    Unauthorized:
      description: Missing, malformed, or expired token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Caller lacks the required role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Request body or query failed validation
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Problem'
              - type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field: { type: string }
                        message: { type: string }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status, detail, instance]
      properties:
        type: { type: string, format: uri, example: 'https://httpstatuses.com/404' }
        title: { type: string, example: Not Found }
        status: { type: integer, example: 404 }
        detail: { type: string, example: Movie not found }
        instance: { type: string, example: /api/movies/65f3... }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, example: Movie deleted successfully }
    Paginated:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          required: [page, limit, total, totalPages]
          properties:
            page: { type: integer, example: 1 }
            limit: { type: integer, example: 10 }
            total: { type: integer, example: 42 }
            totalPages: { type: integer, example: 5 }
    User:
      type: object
      required: [_id, name, email, role, isActive, walletBalance, createdAt, updatedAt]
      properties:
        _id: { type: string, example: 65f3a1d2c8e7b9001fae12d4 }
        name: { type: string, maxLength: 150 }
        email: { type: string, format: email, maxLength: 100 }
        role: { type: string, enum: [user, admin] }
        birthDate: { type: [string, 'null'], format: date-time }
        walletBalance: { type: number, minimum: 0 }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Review:
      type: object
      required: [_id, userId, movieId, comment, rating, createdAt]
      properties:
        _id: { type: string }
        userId:
          description: Populated with `{ _id, email }` on read endpoints, plain ObjectId on create.
          oneOf:
            - type: string
            - type: object
              properties:
                _id: { type: string }
                email: { type: string, format: email }
        movieId:
          description: Populated with `{ _id, title }` on read endpoints, plain ObjectId on create.
          oneOf:
            - type: string
            - type: object
              properties:
                _id: { type: string }
                title: { type: string }
        comment: { type: string, maxLength: 500 }
        rating: { type: number, minimum: 0, maximum: 10 }
        isSpoiler: { type: boolean, default: false }
        createdAt: { type: string, format: date-time }
    ReviewCreateRequest:
      type: object
      required: [movieId, comment, rating]
      properties:
        movieId: { type: string, description: ObjectId of the movie being reviewed }
        comment: { type: string, minLength: 1, maxLength: 500 }
        rating: { type: number, minimum: 0, maximum: 10 }
        isSpoiler: { type: boolean, default: false }
    ReviewUpdateRequest:
      type: object
      properties:
        comment: { type: string, maxLength: 500 }
        rating: { type: number, minimum: 0, maximum: 10 }
        isSpoiler: { type: boolean }